Refactor predicates, add onSendCallback
The predicates were unnecessarily different between the server and
browser. This unifies them as much as possible. We also add a callback
`onSendCallback` with the same signature as `checkIgnore`, this fixes #477.
The point of this callback is meant to separate the concern of ignoring
errors and having a side-channel hook for doing extra things in your app
based on errors that Rollbar is already handling.
